Cleanup IJupyterSession
